Convert coverage-total script from TypeScript to JavaScript
